requirements.txt
# requirements.txt

PyMuPDF
python-docx
nltk
sumy
transformers
torch
sentencepiece
rouge-score